Move Agda.Utils.Geniplate to Agda.Syntax.Geniplate
