+++ src/full/Agda/Syntax/Abstract.hs
@@ -33,6 +33,7 @@ import Agda.Syntax.Abstract.Name
 import Agda.Syntax.Abstract.Name as A (QNamed)
 import qualified Agda.Syntax.Internal as I
 import Agda.Syntax.Common
+import Agda.Syntax.Geniplate
 import Agda.Syntax.Info
 import Agda.Syntax.Literal
 import Agda.Syntax.Position
@@ -40,7 +41,6 @@ import Agda.Syntax.Scope.Base
 
 import Agda.TypeChecking.Positivity.Occurrence
 
-import Agda.Utils.Geniplate
 import Agda.Utils.Lens
 import Agda.Utils.Pretty
